public/voice-engine/video_requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.0.0
torch>=2.1.0
torchvision>=0.16.0
diffusers>=0.27.0
transformers>=4.38.0
accelerate>=0.27.0
imageio>=2.34.0
imageio-ffmpeg>=0.4.9
opencv-python-headless>=4.9.0
